+++ src/mechanisms/hmacblake2s.rs
@@ -60,7 +60,7 @@ impl MechanismImpl for super::HmacBlake2s {
 
         mac.update(&request.message);
         let result = mac.finalize();
-        let signature = Signature::from_slice(&result.into_bytes()).unwrap();
+        let signature = Signature::try_from(&*result.into_bytes()).unwrap();
 
         Ok(reply::Sign { signature })
     }