+++ src/mechanisms/hmacsha1.rs
@@ -61,7 +61,7 @@ impl MechanismImpl for super::HmacSha1 {
 
         mac.update(&request.message);
         let result = mac.finalize();
-        let signature = Signature::from_slice(&result.into_bytes()).unwrap();
+        let signature = Signature::try_from(&*result.into_bytes()).unwrap();
 
         Ok(reply::Sign { signature })
     }