+++ src/mechanisms/hmacsha256.rs
@@ -66,7 +66,7 @@ impl MechanismImpl for super::HmacSha256 {
 
         mac.update(&request.message);
         let result = mac.finalize();
-        let signature = Signature::from_slice(&result.into_bytes()).unwrap();
+        let signature = Signature::try_from(&*result.into_bytes()).unwrap();
 
         Ok(reply::Sign { signature })
     }