+++ core/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
--
+### Changed
+
+- Deprecate `FilesystemClient::debug_dump_store`.  Instead, a debugger should be used to extract the filesystem from a development device.
 
 ## [v0.1.0](https://github.com/trussed-dev/trussed/releases/tag/core-v0.1.0) (2025-01-08)