+++ core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - Deprecate `FilesystemClient::debug_dump_store`.  Instead, a debugger should be used to extract the filesystem from a development device.
+- Update to `heapless-bytes` v0.5.
 
 ## [v0.1.0](https://github.com/trussed-dev/trussed/releases/tag/core-v0.1.0) (2025-01-08)